Add blend tree to new 11 attack animation

+++ KungFu/Assets/CleanIDSmb.cs
@@ -6,7 +6,7 @@ public class CleanIDSmb : BaseSmb
 {
     public override void OnStateEnter(Animator animator, AnimatorStateInfo animatorStateInfo, int layerIndex)
     {
-        animator.SetInteger("AttackID_i", -1);
+        animator.SetFloat("AttackID_i", -1);
         animator.SetInteger("DefenseID_i", -1);
     }
 }
